// src/common/enums/user-role.enum.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  SALES_REP = 'SALES_REP',
  SALES_MANAGER = 'SALES_MANAGER',
  WAREHOUSE = 'WAREHOUSE',
  WAREHOUSE_KEEPER = 'WAREHOUSE_KEEPER',
  WH_MANAGER = 'WH_MANAGER',
  WAREHOUSE_MANAGER = 'WAREHOUSE_MANAGER',
  ACCOUNTANT = 'ACCOUNTANT',
  CUSTOMER = 'CUSTOMER',
}